Add ability to reduce snake width in code [Snake Game]
Significant changes to the snake's movement logic. Reducing the width of the snake required the development of turning logic to correctly represent this movement. Reducing the width contributed to unwanted parts of the snake or gaps due to the snake going through the 'walls of the board'. Removed the grid from gridLayer.